Update to JRE Version 8 Update 161 released January 16, 2018
Due to Oracle using unpack2000 for the RPM installation which is an
ELF32 binary, this build now installs glibc.i686 and libgcc.i686 so that the
rpm installation doesn't fail. Once the RPM is installed, the two packages, and
a single dependent are uninstalled maintaining the size of the image.